+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Version 1.5-SNAPSHOT
 
 ADDED:
 
-- new JToolbarItemInfo class created for build JButton objects from given fields for add on a JToolBar object
+- new factory methods for JMenu and JMenuItem with a JMenuItemInfo as argument
 
 Version 1.4
 -------------